third_party/blink/web_tests/external/wpt/content-security-policy/gen/top.http-rp/script-src-self/worker-import.http-expected.txt

This is a testharness.js-based test.
PASS Content Security Policy: Expects allowed for worker-import to same-http origin and keep-origin redirection from http context.
PASS Content Security Policy: Expects allowed for worker-import to same-http origin and keep-origin redirection from http context.: securitypolicyviolation
PASS Content Security Policy: Expects allowed for worker-import to same-http origin and no-redirect redirection from http context.
PASS Content Security Policy: Expects allowed for worker-import to same-http origin and no-redirect redirection from http context.: securitypolicyviolation
PASS Content Security Policy: Expects blocked for worker-import to cross-http origin and keep-origin redirection from http context.
FAIL Content Security Policy: Expects blocked for worker-import to cross-http origin and keep-origin redirection from http context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
PASS Content Security Policy: Expects blocked for worker-import to cross-http origin and no-redirect redirection from http context.
FAIL Content Security Policy: Expects blocked for worker-import to cross-http origin and no-redirect redirection from http context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
PASS Content Security Policy: Expects blocked for worker-import to cross-http origin and swap-origin redirection from http context.
FAIL Content Security Policy: Expects blocked for worker-import to cross-http origin and swap-origin redirection from http context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
PASS Content Security Policy: Expects blocked for worker-import to same-http origin and swap-origin redirection from http context.
FAIL Content Security Policy: Expects blocked for worker-import to same-http origin and swap-origin redirection from http context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
Harness: the test ran to completion.
